refactor: module-app 이름을 module-domain 으로 변경한다
