+++ order/order-repository/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    implementation project(':order:order-app')
+    implementation project(':order:order-domain')
 
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
