+++ settings.gradle
@@ -15,7 +15,7 @@ include 'support'
 include 'support:jacoco'
 
 include 'shop'
-include 'shop:shop-app'
+include 'shop:shop-domain'
 include 'shop:shop-service'
 include 'shop:shop-controller'
 include 'shop:shop-repository'
@@ -28,13 +28,9 @@ include 'auth:auth-repository'
 
 include 'core'
 include 'core:id-gen'
-include 'order'
-include 'order:order-app'
-findProject(':order:order-app')?.name = 'order-app'
-include 'order:order-controller'
-findProject(':order:order-controller')?.name = 'order-controller'
-include 'order:order-service'
-findProject(':order:order-service')?.name = 'order-service'
-include 'order:order-repository'
-findProject(':order:order-repository')?.name = 'order-repository'
 
+include 'order'
+include 'order:order-domain'
+include 'order:order-service'
+include 'order:order-controller'
+include 'order:order-repository'
