+++ shop/shop-domain/src/main/java/shop/woowasap/shop/app/api/request/UpdateProductRequest.java
