+++ shop/shop-domain/src/main/java/shop/woowasap/shop/app/exception/InvalidProductDescriptionException.java
