+++ shop/shop-repository/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    implementation project(':shop:shop-app')
+    implementation project(':shop:shop-domain')
 
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
