+++ shop/shop-service/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
     implementation project(':core:id-gen')
-    implementation project(':shop:shop-app')
+    implementation project(':shop:shop-domain')
 
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 }
