+++ shop/shop-service/src/main/java/shop/woowasap/shop/service/mapper/ProductMapper.java
@@ -2,7 +2,7 @@ package shop.woowasap.shop.service.mapper;
 
 import java.time.ZoneOffset;
 import shop.woowasap.core.id.api.IdGenerator;
-import shop.woowasap.shop.app.domain.product.Product;
+import shop.woowasap.shop.app.product.Product;
 import shop.woowasap.shop.app.api.request.RegisterProductRequest;
 
 public final class ProductMapper {
